+++ tests/settings.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import os
 
 DATABASES = {
     'default': {
@@ -17,11 +16,6 @@ INSTALLED_APPS = [
     'tests.test_us',
     'tests.test_pk',
 ]
-
-import django
-
-if django.VERSION[:2] < (1, 6):
-    TEST_RUNNER = 'discover_runner.DiscoverRunner'
 
 SECRET_KEY = 'spam-spam-spam-spam'